feat: refactor shared functions to common and update package structure
