+++ services/auth-service/pom.xml
@@ -3,12 +3,11 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.5.4</version>
-        <relativePath/> <!-- lookup parent from repository -->
+        <groupId>com.finedine</groupId>
+        <artifactId>finedine-parent</artifactId>
+        <version>0.0.1-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>  <!-- Points to root POM -->
     </parent>
-    <groupId>com.finedine</groupId>
     <artifactId>auth-service</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>auth-service</name>
@@ -30,6 +29,7 @@
         <java.version>21</java.version>
         <bucket4j.version>8.10.1</bucket4j.version>
         <spring-cloud.version>2025.0.0</spring-cloud.version>
+        <version>0.0.1-SNAPSHOT</version>
     </properties>
     <dependencies>
         <dependency>
@@ -142,7 +142,6 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-data-redis</artifactId>
         </dependency>
-        <!-- Rate Limiting Dependencies -->
         <dependency>
             <groupId>com.bucket4j</groupId>
             <artifactId>bucket4j-core</artifactId>
@@ -170,6 +169,11 @@
             <groupId>org.springframework.cloud</groupId>
             <artifactId>spring-cloud-starter-netflix-eureka-client</artifactId>
             <version> 4.3.0</version>
+        </dependency>
+        <dependency>
+            <groupId>com.finedine</groupId>
+            <artifactId>common</artifactId>
+            <version>${project.version}</version>
         </dependency>
     </dependencies>
     <dependencyManagement>
