+++ services/auth-service/src/main/java/com/finedine/authservice/AuthServiceApplication.java
@@ -1,9 +1,10 @@
 package com.finedine.authservice;
 
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = "com.finedine")
 public class AuthServiceApplication {
 
     public static void main(String[] args) {
