+++ services/auth-service/src/main/java/com/finedine/authservice/dto/signup/CustomerRegistrationRequest.java
@@ -1,6 +1,7 @@
 package com.finedine.authservice.dto.signup;
 
-import com.finedine.authservice.util.contraints.ImageFiles;
+//import com.finedine.authservice.util.contraints.ImageFiles;
+import com.finedine.common.validators.ImageFiles;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
 import org.springframework.web.multipart.MultipartFile;
