+++ services/auth-service/src/main/java/com/finedine/authservice/dto/signup/RestaurantRegistrationRequest.java
@@ -1,7 +1,7 @@
 package com.finedine.authservice.dto.signup;
 
 
-import com.finedine.authservice.util.contraints.ImageFiles;
+import com.finedine.common.validators.ImageFiles;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
