+++ services/auth-service/src/main/java/com/finedine/authservice/dto/signup/RiderRegistrationRequest.java
@@ -1,7 +1,7 @@
 package com.finedine.authservice.dto.signup;
 
 import com.finedine.authservice.enums.VehicleType;
-import com.finedine.authservice.util.contraints.ImageFiles;
+import com.finedine.common.validators.ImageFiles;
 
 import jakarta.validation.constraints.NotBlank;
 import org.springframework.web.multipart.MultipartFile;
