--- services/auth-service/src/main/java/com/finedine/authservice/exception/GlobalExceptionHandler.java
@@ -1,139 +1,0 @@
-package com.finedine.authservice.exception;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ProblemDetail;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.client.HttpClientErrorException;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import static com.finedine.authservice.CustomMessages.*;
-
-@Slf4j
-@ControllerAdvice
-@RestController
-public class GlobalExceptionHandler {
-
-    private static final String timestamp = "timestamp";
-    //todo: add error urls to problem detail
-
-
-    @ExceptionHandler(VerificationFailedException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ProblemDetail handleVerificationFailedException(VerificationFailedException e) {
-
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.BAD_REQUEST, e.getMessage());
-        problemDetail.setTitle("Verification Failed");
-        problemDetail.setProperty(timestamp, LocalDateTime.now());
-        return problemDetail;
-    }
-
-    @ExceptionHandler(BadCredentialsException.class)
-    @ResponseStatus(HttpStatus.UNAUTHORIZED)
-    public ProblemDetail handleBadCredentialsException(BadCredentialsException e) {
-        log.error("Bad credentials.", e);
-
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.UNAUTHORIZED, e.getMessage());
-
-        problemDetail.setTitle("Unauthorized");
-        problemDetail.setProperty(timestamp, LocalDateTime.now());
-        return problemDetail;
-    }
-
-    @ExceptionHandler(Exception.class)
-    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-    public ProblemDetail handleGenericException(Exception e) {
-        log.error("Internal Server Error occurred.", e);
-
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
-
-        problemDetail.setTitle("Internal Server Error");
-        problemDetail.setProperty(timestamp, LocalDateTime.now());
-        return problemDetail;
-    }
-
-
-    @ExceptionHandler(NotFoundException.class)
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ProblemDetail handleNotFoundException(NotFoundException e) {
-        log.error("NotFoundException occurred.", e);
-
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.NOT_FOUND, e.getMessage());
-        problemDetail.setTitle("Resource Not Found");
-        problemDetail.setProperty(timestamp, LocalDateTime.now());
-        return problemDetail;
-    }
-
-
-    @ExceptionHandler(UnauthorizedException.class)
-    @ResponseStatus(HttpStatus.UNAUTHORIZED)
-    public ProblemDetail handleUnauthorizedException(UnauthorizedException e) {
-        log.error("Unauthorized access exception occurred.", e);
-
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.UNAUTHORIZED, e.getMessage());
-        problemDetail.setTitle("Unauthorized Access");
-        problemDetail.setProperty(timestamp, LocalDateTime.now());
-        return problemDetail;
-    }
-
-
-    @ExceptionHandler(UnverifiedAccountException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ProblemDetail handleUserNotVerifiedException(UnverifiedAccountException e) {
-        log.error("UnverifiedAccountException occurred.", e);
-
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.BAD_REQUEST, e.getMessage());
-        problemDetail.setTitle(USER_NOT_VERIFIED_MSG);
-        problemDetail.setProperty(timestamp, LocalDateTime.now());
-        return problemDetail;
-    }
-
-
-    @ExceptionHandler(HttpClientErrorException.BadRequest.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ProblemDetail handleBadRequestException(Exception e) {
-        log.error("BadRequestException occurred.", e);
-
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.BAD_REQUEST, e.getMessage());
-        problemDetail.setTitle("Bad Request");
-        problemDetail.setProperty(timestamp, LocalDateTime.now());
-        return problemDetail;
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ProblemDetail handleIllegalArgumentException(IllegalArgumentException e) {
-        log.error("Validation failed: {}", e.getMessage());
-
-        ProblemDetail pd = ProblemDetail.forStatusAndDetail(HttpStatus.BAD_REQUEST, e.getMessage());
-        pd.setTitle("Validation Error");
-        pd.setProperty(timestamp, LocalDateTime.now());
-        return pd;
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ProblemDetail handleMethodArgumentNotValidException(MethodArgumentNotValidException e) {
-        log.error("Validation failed: {}", e.getMessage());
-        List<String> errors = e.getBindingResult().getFieldErrors().stream()
-                .map(error -> error.getField() + ": " + error.getDefaultMessage())
-                .collect(Collectors.toList());
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.BAD_REQUEST, "Validation failed for request");
-        problemDetail.setTitle("Invalid Request");
-        problemDetail.setProperty(timestamp, LocalDateTime.now());
-        problemDetail.setProperty("errors", errors);
-        return problemDetail;
-    }
-
-    public void logException(Exception e) {
-        log.error("Exception occurred. ", e);
-    }
-}
