+++ services/auth-service/src/main/java/com/finedine/authservice/service/AuthServiceImpl.java
@@ -12,14 +12,14 @@ import com.finedine.authservice.entity.AuthToken;
 import com.finedine.authservice.entity.OtpType;
 import com.finedine.authservice.enums.AccountStatus;
 import com.finedine.authservice.enums.Role;
-import com.finedine.authservice.exception.NotFoundException;
-import com.finedine.authservice.exception.VerificationFailedException;
+import com.finedine.common.exception.NotFoundException;
+import com.finedine.common.exception.VerificationFailedException;
 import com.finedine.authservice.repository.AccountRepository;
 import com.finedine.authservice.security.SecurityUser;
 import com.finedine.authservice.security.TokenManager;
 import com.finedine.authservice.util.Common;
 import com.finedine.authservice.util.config.SqsProducer;
-import com.finedine.authservice.util.contraints.CustomValidator;
+import com.finedine.common.validators.CustomValidator;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
