+++ services/auth-service/src/main/java/com/finedine/authservice/service/RefreshTokenService.java
@@ -3,8 +3,8 @@ package com.finedine.authservice.service;
 import com.finedine.authservice.entity.Account;
 import com.finedine.authservice.entity.AuthToken;
 import com.finedine.authservice.entity.RefreshToken;
-import com.finedine.authservice.exception.NotFoundException;
-import com.finedine.authservice.exception.UnauthorizedException;
+import com.finedine.common.exception.NotFoundException;
+import com.finedine.common.exception.UnauthorizedException;
 import com.finedine.authservice.repository.AccountRepository;
 import com.finedine.authservice.repository.RefreshTokenRepository;
 import com.finedine.authservice.security.TokenManager;
