+++ services/common/src/main/java/com/finedine/common/config/CommonConfig.java
@@ -1,0 +1,4 @@
+package com.finedine.common.config;
+
+public class CommonConfig {
+}
