+++ services/common/src/main/java/com/finedine/common/exception/NotFoundException.java
@@ -1,5 +1,4 @@
-package com.finedine.authservice.exception;
-
+package com.finedine.common.exception;
 public class NotFoundException extends RuntimeException {
     public NotFoundException(String message) {
         super(message);
