+++ services/common/src/main/java/com/finedine/common/exception/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package com.finedine.authservice.exception;
+package com.finedine.common.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
