+++ services/common/src/main/java/com/finedine/common/exception/UnverifiedAccountException.java
@@ -1,5 +1,4 @@
-package com.finedine.authservice.exception;
-
+package com.finedine.common.exception;
 public class UnverifiedAccountException extends RuntimeException {
     public UnverifiedAccountException(String message) {
         super(message);
