+++ services/common/src/main/java/com/finedine/common/validators/ImageFilesValidator.java
@@ -1,4 +1,4 @@
-package com.finedine.authservice.util.contraints;
+package com.finedine.common.validators;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
