+++ services/restaurant-service/src/main/java/com/finedine/restaurantservice/RestaurantServiceApplication.java
@@ -3,7 +3,7 @@ package com.finedine.restaurantservice;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = "com.finedine")
 public class RestaurantServiceApplication {
 
     public static void main(String[] args) {
