--- services/restaurant-service/src/main/java/com/finedine/restaurantservice/exception/GlobalExceptionHandler.java
@@ -1,46 +1,0 @@
-package com.finedine.restaurantservice.exception;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ProblemDetail;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.time.LocalDateTime;
-
-import static com.finedine.restaurantservice.util.CustomMessages.RESTAURANT_NOT_FOUND;
-import static com.finedine.restaurantservice.util.CustomMessages.RESTRICTED_ACTION;
-
-@Slf4j
-@ControllerAdvice
-@RestController
-public class GlobalExceptionHandler {
-     private static String timestamp = "Timestamp";
-
-    @ExceptionHandler(NotFoundException.class)
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ProblemDetail handleNotFoundException(NotFoundException e) {
-        log.error("NotFoundException occurred.", e);
-
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.NOT_FOUND, RESTAURANT_NOT_FOUND);
-        problemDetail.setTitle("Resource Not Found");
-        problemDetail.setProperty(timestamp, LocalDateTime.now());
-        return problemDetail;
-    }
-
-
-    @ExceptionHandler({UnauthorizedException.class, AccessDeniedException.class})
-    @ResponseStatus(HttpStatus.UNAUTHORIZED)
-    public ProblemDetail handleUnauthorizedException(UnauthorizedException e) {
-        log.error("Unauthorized access exception occurred.", e);
-
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.UNAUTHORIZED, RESTRICTED_ACTION);
-        problemDetail.setTitle("Unauthorized Access");
-        problemDetail.setProperty(timestamp, LocalDateTime.now());
-        return problemDetail;
-    }
-
-}
