+++ services/restaurant-service/src/main/java/com/finedine/restaurantservice/service/RestaurantServiceImpl.java
@@ -3,8 +3,8 @@ package com.finedine.restaurantservice.service;
 import com.finedine.restaurantservice.dto.*;
 import com.finedine.restaurantservice.entity.MenuItem;
 import com.finedine.restaurantservice.entity.Restaurant;
-import com.finedine.restaurantservice.exception.NotFoundException;
-import com.finedine.restaurantservice.exception.UnauthorizedException;
+import com.finedine.common.exception.NotFoundException;
+import com.finedine.common.exception.UnauthorizedException;
 import com.finedine.restaurantservice.repository.MenuRepository;
 import com.finedine.restaurantservice.repository.RestaurantRepository;
 import com.finedine.restaurantservice.security.SecurityUser;
